Refactor Duke class to NyanDuke class
