--- src/main/java/duke/gui/Main.java
@@ -1,36 +1,0 @@
-package duke.gui;
-
-import java.io.IOException;
-
-import duke.Duke;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
-/**
- * A GUI for Duke using FXML.
- */
-public class Main extends Application {
-    private Duke duke = new Duke();
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            MainWindow controller = fxmlLoader.getController();
-            controller.setDuke(duke);
-            controller.showWelcome();
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
