+++ src/main/java/nyanduke/gui/Main.java
@@ -1,0 +1,36 @@
+package nyanduke.gui;
+
+import java.io.IOException;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+import nyanduke.NyanDuke;
+
+/**
+ * A GUI for Duke using FXML.
+ */
+public class Main extends Application {
+    private NyanDuke nyanDuke = new NyanDuke();
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            MainWindow controller = fxmlLoader.getController();
+            controller.setNyanDuke(nyanDuke);
+            controller.showWelcome();
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
